requirements.txt
solana==0.30.2
solders==0.18.1
python-dotenv==1.0.1
rich==13.9.3
typer==0.12.5
pydantic==2.9.2
httpx==0.23.3
click==8.1.7
